auth_service/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
redis
django-redis
